+++ great_expectations/core/usage_statistics/usage_statistics_record_schema.json
@@ -977,7 +977,8 @@
         "version": {
           "enum": [
             "1.0.0",
-            "1.0.1"
+            "1.0.1",
+            "1.0.2"
           ]
         },
         "platform.system": {
@@ -5195,7 +5196,8 @@
     "version": {
       "enum": [
         "1.0.0",
-        "1.0.1"
+        "1.0.1",
+        "1.0.2"
       ]
     },
     "ge_version": {
